blueprint cinema: genres and movies, relationship + CRUDs

+++ app/auth/routes.py
@@ -3,11 +3,11 @@ from app.auth import auth_bp as auth
 
 from flask import render_template, redirect, url_for, flash, request
 from flask_login import login_user, logout_user, login_required, current_user
-from wtforms.validators import DataRequired
-
+from wtforms.validators import DataRequired, Optional
+from sqlalchemy.exc import IntegrityError
 
 from app import db
-from app.auth.forms import RegisterNewUserForm, RegisterRoleForm
+from app.auth.forms import RegisterNewUserForm, RegisterRoleForm, LoginUserForm
 from app.models.auth_user import User
 from app.models.auth_role import Role
 
@@ -15,9 +15,9 @@ from app.models.auth_role import Role
 # Create
 @auth.route('/user/role/create', methods=['GET', 'POST'])
 def register_role():
-    if current_user.is_authenticated:
-        flash('You are already logged in.', 'info')
-        return redirect(url_for('cinema.index'))
+    # if current_user.is_authenticated:
+    #     flash('You are already logged in.', 'info')
+    #     return redirect(url_for('cinema.index'))
     
     form = RegisterRoleForm()
 
@@ -26,9 +26,16 @@ def register_role():
             name = form.name.data,
             description = form.description.data
         )
-        db.session.add(new_role)
-        db.session.commit()
-        flash('Role Created', 'alert')
+        try:
+            db.session.add(new_role)
+            db.session.commit()
+            flash('Role Created', 'alert')
+        except IntegrityError:
+            db.session.rollback()
+            flash('The role could not be created: it already exists or there is an integrity conflict.', 'error')
+        except Exception as e:
+            db.session.rollback()
+            flash('Unexpected error: {str(e)}', 'error')
         return redirect( url_for('auth.view_roles') )
     return render_template('auth_role_create.html', form=form)
 
@@ -39,7 +46,7 @@ def view_roles():
         db.select(Role).order_by(Role.id.asc())
     ).scalars().all()
 
-    return render_template('auth_role_list.html', roles=roles)
+    return render_template("auth_role_list.html", roles=roles)
 
 # Update
 @auth.route('/user/role/<int:role_id>', methods=['GET', 'POST'])
@@ -54,12 +61,18 @@ def update_role(role_id):
         description = role.description
     )
 
+    form.submit.label.text = 'Update'
+
     if form.validate_on_submit():
         role.name = form.name.data
         role.description = form.description.data
 
-        db.session.commit()
-        flash("Updated role", "success")
+        try:
+            db.session.commit()
+            flash("Updated role", "success")
+        except:
+            db.session.rollback()
+            flash("The role could not be updated due to integrity restrictions.", "error")
         return redirect( url_for('auth.view_roles') )
     
     elif request.method == 'GET':
@@ -73,27 +86,34 @@ def update_role(role_id):
 def delete_role(role_id):
     role = db.session.get(Role, role_id)
     if not role:
-        flash("Role doesn't exist", "")
+        flash("Role doesn't exist", "alert")
         return redirect( url_for('auth.view_roles') )
     
-    db.session.delete(role)
-    db.session.commit()
-    flash("Deleted role", "")
+    if User.query.filter_by(role_id=role_id).count() > 0:
+        flash("The role in use cannot be deleted.")
+    else:
+        try:
+            db.session.delete(role)
+            db.session.commit()
+            flash("Deleted role", "")
+        except:
+            db.session.rollback()
+            flash("The role cannot be deleted due to integrity constraints.")
     return redirect( url_for('auth.view_roles') )
 
 ## USERS
 # CREATE
 @auth.route('/user/create', methods=['GET', 'POST'])
 def register_user():
-    if current_user.is_authenticated:
-        flash('You are already logged in.', 'info')
-        return redirect(url_for('cinema.index'))
+    # if current_user.is_authenticated:
+    #     flash('You are already logged in.', 'info')
+    #     return redirect(url_for('cinema.index'))
     
     form = RegisterNewUserForm()
 
     roles = Role.query.all()
     form.password.validators = [DataRequired()]
-    form.role.choices = [('2', '-- Default --')] + [(r.id, r.name) for r in roles]
+    form.role.choices = [(2, '-- Default --')] + [(r.id, r.name) for r in roles]
     
     if form.validate_on_submit():
         new_user = User.create_user(
@@ -103,9 +123,16 @@ def register_user():
             role_id = form.role.data
         )
 
-        db.session.add(new_user)
-        db.session.commit()
-        flash('register successful')
+        try:
+            db.session.add(new_user)
+            db.session.commit()
+            flash('register successful')
+        except IntegrityError:
+            db.session.rollback()
+            flash("The user could not be created: it already exists or there is an integrity conflict.", "error")
+        except Exception as e:
+            db.session.rollback()
+            flash("Unexpected error: {str(e)}", "error")
         return redirect( url_for('auth.view_users') )
 
     return render_template('auth_user_create.html', form=form)
@@ -130,6 +157,7 @@ def update_user(user_id):
     form = RegisterNewUserForm(user=user)
     form.password.validators = []
     form.role.choices = [(user.role_id, f'-- { user.roles.name } --')] + [(role.id, role.name) for role in Role.query.all()]
+    form.submit.label.text = "Update"
     
     if form.validate_on_submit():
         user.name = form.name.data
@@ -139,14 +167,17 @@ def update_user(user_id):
         if form.password.data:
             user.password_hash = form.password.data
 
-        db.session.commit()
-        flash("User updated successfully", "")
+        try:
+            db.session.commit()
+            flash("User updated successfully", "")
+        except:
+            db.session.rollback()
+            flash("The user could not be updated due to integrity restrictions.", "")
         return redirect( url_for('auth.view_users') )
     elif request.method == 'GET':
         form.name.data = user.name
         form.email.data = user.email
         form.role.data = user.role_id
-
 
     return render_template('auth_user_create.html', form=form)
 
@@ -159,15 +190,69 @@ def delete_user(user_id):
         flash("User doesn't exist", "")
         return redirect( url_for('view_users') )
     
-    db.session.delete(user)
-    db.session.commit()
-    flash("Deleted user", "")
+    try:
+        db.session.delete(user)
+        db.session.commit()
+        flash("Deleted user", "")
+    except:
+        db.session.rollback()
+        flash("The user cannot be deleted due to integrity constraints.", "")
     return redirect( url_for('auth.view_users') )
 #
 
-@auth.route('/login')
+@auth.route('/login', methods=['GET', 'POST'])
 def login():
-    pass
+    form = LoginUserForm()
+
+    if current_user.is_authenticated:
+        flash("You already have an active session.", "alert")
+
+    if form.validate_on_submit():
+        user = db.session.execute(db.select(User).filter(
+            User.email == form.email.data
+        )).scalar_one_or_none()
+
+        if not user or not user.check_password(form.password.data):
+            flash("Datos invalidaos", "alert-danger")
+            return redirect( url_for("auth.login") )
+        
+        login_user(user, form.stay_loggedin.data)
+
+        next_page = request.args.get('next')
+
+        flash(f"Welcome {user.name.capitalize()}", "alert-success")
+        return redirect( next_page or url_for('cinema.index') )
+
+    return render_template('auth_login.html', form=form)
+
+@auth.route('/register', methods=['GET', 'POST'])
+def register():
+    form = RegisterNewUserForm()
+
+    form.role.choices = [(2, 'User')]
+    form.role.validators = [Optional()]
+
+    if form.validate_on_submit():
+        new_user = User.create_user(
+            name = form.name.data,
+            email = form.email.data,
+            password = form.password.data,
+            role_id = 2
+        )
+        
+        try:
+            db.session.add(new_user)
+            db.session.commit()
+            flash('register successful')
+        except IntegrityError:
+            db.session.rollback()
+            flash("The user could not be created: it already exists or there is an integrity conflict.", "error")
+        except Exception as e:
+            db.session.rollback()
+            flash("Unexpected error: {str(e)}", "error")
+        return redirect( url_for('auth.login') )
+
+    return render_template('auth_register.html', form=form)
 
 @auth.route('/logout')
 def logout():
